feat(chi): replacing mux with chi router
