+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/ClemSK/chirpy
 
 go 1.21.3
+
+require github.com/go-chi/chi/v5 v5.0.10 // indirect
